call api on map is added but it doesnt work yet
